- It's now possible to use random block palettes for streets too - Big speed improvement in the Lost City profile selection screen when rendering the world preview - Saplings in Lost City parks will generate trees again - 'cityChance' now has double precision which makes it possible to make cities even more rare - More expansions to the lost city api

+++ src/main/java/mcjty/lostcities/worldgen/lost/City.java
@@ -89,15 +89,15 @@ public class City {
             CitySphere sphere = CitySphere.getCitySphere(chunkX, chunkZ, provider);
             if (!sphere.isEnabled()) {
                 // No sphere
-                return rand.nextFloat() < provider.getOutsideProfile().CITY_CHANCE;
+                return rand.nextDouble() < provider.getOutsideProfile().CITY_CHANCE;
             }
             if (sphere.getCenter().chunkX() == chunkX && sphere.getCenter().chunkZ() == chunkZ) {
                 // This chunk is the center of a city
-                return rand.nextFloat() < provider.getProfile().CITY_CHANCE;
+                return rand.nextDouble() < provider.getProfile().CITY_CHANCE;
             }
             return false;
         } else {
-            return rand.nextFloat() < provider.getProfile().CITY_CHANCE;
+            return rand.nextDouble() < provider.getProfile().CITY_CHANCE;
         }
     }
